Replace pudl_out scope=module fixture with a scope=session fixture defined in conftest. Use the one from output_test that is fast_out_annual for faster tests
